Allow the user to provide CAS digests as inputs
This commit allows for a new kind of input: InputDigests. This is a map from
the abosolute path of a file to it's digest. A user can provide this output that
was returned from a previous run - for example. To avoid having to download and
upload files (or having to check if the files were updated before).